Moved service methods from LikeService to PostService. Added missing where arg to query. new migrations
