+++ src/Cobase/AppBundle/Service/LikeService.php
@@ -40,45 +40,8 @@ class LikeService
      *
      * @return Like
      */
-    public function likePost(Post $post, User $user)
+    public function getPostSpecificLikeByUser(Post $post, User $user)
     {
-        if ($user->likesPost($post)) {
-            throw new Exception('You already like this post');
-        }
-
-        $like = new Like($post);
-
-        $like->setUser($user);
-
-        $this->em->getConnection()->beginTransaction();
-        $this->em->persist($like);
-        $this->em->flush();
-
-        $this->em->getConnection()->commit();
-
-        return $like;
-    }
-
-    /**
-     * @param Post $post
-     * @param User $user
-     * @throws \Exception
-     */
-    public function unlikePost(Post $post, User $user)
-    {
-        if (!$user->likesPost($post)) {
-            throw new Exception("You don't like this post");
-        }
-
-        $like = $this->likeRepository->getPostSpecificLikeByUser($post, $user);
-
-        if (null === $like) {
-            throw new ResourceNotFoundException('No like found for this post');
-        }
-
-        $this->em->getConnection()->beginTransaction();
-        $this->em->remove($like);
-        $this->em->flush();
-        $this->em->getConnection()->commit();
+        return $this->likeRepository->getPostSpecificLikeByUser($post, $user);
     }
 }
